pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-my-docs"
version = "0.1.0"
description = "Domain-specific RAG over Project Gutenberg history & philosophy texts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",
    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    # Ingestion
    "requests>=2.32.0",
    "nltk>=3.8.1",
    "spacy>=3.7.0",
    "tiktoken>=0.7.0",
    "jsonlines>=4.0.0",
    # Indexing & Retrieval
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "qdrant-client>=1.10.0",
    # Generation
    "openai>=1.35.0",
    "anthropic>=0.30.0",
    # Evaluation
    "ragas>=0.1.14",
    "datasets>=2.20.0",
    # Observability
    "structlog>=24.2.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-requests>=2.32.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
amd = "amd.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/amd"]

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["amd"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
files = ["src/amd"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=amd --cov-report=term-missing"